% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annot.R
\name{create_annot}
\alias{create_annot}
\title{Annotation object creation and alteration}
\usage{
create_annot(
  mat = NULL,
  mat_file = NULL,
  mat_dir = getwd(),
  mat_import_method = NULL,
  format_matrix = TRUE,
  convert_to = "double",
  rownames.force = NA,
  rownames = NULL,
  colnames = NULL,
  mat_import_opts = NULL,
  annot_type = NULL,
  annot_obj = NULL,
  annot_file = NULL,
  annot_dir = getwd(),
  annot_import_method = NULL,
  annot_import_opts = NULL,
  annot_join = NULL,
  remove_missing = FALSE,
  IDs = NULL,
  pseudotime_labs = NULL,
  remove_labs_pattern = NULL,
  drop_labs_containing = NULL,
  select_labs_containing = NULL,
  pseudotimes = NULL,
  remove_pseudotimes_pattern = NULL,
  drop_pseudotimes_containing = NULL,
  select_pseudotimes_containing = NULL,
  color_by = NULL,
  color_palette = NULL,
  select_annot_cols = c(FALSE, FALSE, FALSE, FALSE),
  preserve_names = c(FALSE, FALSE, FALSE, FALSE),
  annot_col_names = c("ID", "Pseudotime Labels", "Pseudotimes", "Pseudotime Colors"),
  replace_annot_cols = c(FALSE, FALSE, FALSE),
  return_as = "default",
  suppress_warnings = FALSE,
  ...
)
}
\arguments{
\item{mat}{Used to specify a matrix object with dimension names. If you have a matrix object in R that contains the dimension names you'd like to use for the new annotation object's IDs, you may provide the object name here. See  Use 'mat' to specify a matrix object in R, or 'mat_file' to specify the name of a file containing a matrix with dimension names for importing. See Details section for more information.}

\item{mat_file, mat_dir}{If you wish to specify}
}
\description{
'create_annot' enables the creation and modification of annotation objects.
}
\details{
The purpose of this function is to output a new annotation object. New objects can be created by supplying a matrix with appropriate dimension names, a pre-existing annotation file, both, or from scratch using this function.
}
